refactor(controller): isolate pool actor lease helpers
