+++ internal/controller/task_controller.go
@@ -79,13 +79,6 @@ const (
 	managedLabelValue      = scheduledRunLabelValue
 
 	workerRBACReconcileFailedReason = "WorkerRBACReconcileFailed"
-
-	substratePoolActorLeasePurpose        = "substrate-actor-pool-lease"
-	substratePoolActorLeaseActorIDLabel   = "orka.ai/substrate-pool-actor-id"
-	substratePoolActorLeaseHolderUIDLabel = "orka.ai/substrate-pool-holder-uid"
-	substratePoolActorLeaseTaskNSAnno     = "orka.ai/substrate-pool-task-namespace"
-	substratePoolActorLeaseTaskNameAnno   = "orka.ai/substrate-pool-task-name"
-	substratePoolActorLeaseTaskUIDAnno    = "orka.ai/substrate-pool-task-uid"
 )
 
 // TaskReconciler reconciles a Task object
@@ -1160,61 +1153,6 @@ func (r *TaskReconciler) tryReserveSubstratePoolActor(
 	return true, nil
 }
 
-func substratePoolActorLeaseHasActiveHolder(ctx context.Context, reader client.Reader, lease *coordinationv1.Lease) (bool, error) {
-	if lease == nil || lease.Annotations == nil {
-		return true, nil
-	}
-	taskNamespace := strings.TrimSpace(lease.Annotations[substratePoolActorLeaseTaskNSAnno])
-	taskName := strings.TrimSpace(lease.Annotations[substratePoolActorLeaseTaskNameAnno])
-	taskUID := strings.TrimSpace(lease.Annotations[substratePoolActorLeaseTaskUIDAnno])
-	if taskNamespace == "" || taskName == "" || taskUID == "" {
-		toolNamespace := strings.TrimSpace(lease.Annotations[substratePoolActorLeaseToolNSAnno])
-		toolName := strings.TrimSpace(lease.Annotations[substratePoolActorLeaseToolNameAnno])
-		if toolNamespace != "" && toolName != "" {
-			return substratePoolActorLeaseHasActiveToolHolder(ctx, reader, toolNamespace, toolName, lease.Annotations[substratePoolActorLeaseToolUIDAnno])
-		}
-		return true, nil
-	}
-	task := &corev1alpha1.Task{}
-	if err := reader.Get(ctx, types.NamespacedName{Namespace: taskNamespace, Name: taskName}, task); err != nil {
-		if apierrors.IsNotFound(err) {
-			return true, nil
-		}
-		return true, err
-	}
-	if string(task.UID) != taskUID {
-		return true, nil
-	}
-	if !task.DeletionTimestamp.IsZero() {
-		return true, nil
-	}
-	if substratePoolActorLeaseActiveTaskPhase(task.Status.Phase) {
-		return true, nil
-	}
-	return taskSubstratePoolActorCleanupRequired(task), nil
-}
-
-func substratePoolActorLeaseHasActiveToolHolder(
-	ctx context.Context,
-	reader client.Reader,
-	toolNamespace string,
-	toolName string,
-	toolUID string,
-) (bool, error) {
-	tool := &corev1alpha1.Tool{}
-	if err := reader.Get(ctx, types.NamespacedName{Namespace: toolNamespace, Name: toolName}, tool); err != nil {
-		if apierrors.IsNotFound(err) {
-			return true, nil
-		}
-		return true, err
-	}
-	toolUID = strings.TrimSpace(toolUID)
-	if toolUID != "" && string(tool.UID) != toolUID {
-		return true, nil
-	}
-	return true, nil
-}
-
 func (r *TaskReconciler) releaseSubstratePoolActorLeases(ctx context.Context, task *corev1alpha1.Task) error {
 	leases, err := r.substratePoolActorLeasesForTask(ctx, task)
 	if err != nil {
@@ -1252,44 +1190,6 @@ func (r *TaskReconciler) deleteSubstratePoolActorLeasesForTask(ctx context.Conte
 		}
 	}
 	return nil
-}
-
-func deleteCurrentObjectPreconditions(obj client.Object) []client.DeleteOption {
-	if obj == nil {
-		return nil
-	}
-	preconditions := &metav1.Preconditions{}
-	if obj.GetUID() != "" {
-		uid := obj.GetUID()
-		preconditions.UID = &uid
-	}
-	if obj.GetResourceVersion() != "" {
-		resourceVersion := obj.GetResourceVersion()
-		preconditions.ResourceVersion = &resourceVersion
-	}
-	if preconditions.UID == nil && preconditions.ResourceVersion == nil {
-		return nil
-	}
-	return []client.DeleteOption{&client.DeleteOptions{Preconditions: preconditions}}
-}
-
-func substrateLeaseStillMatchesAfterDeleteConflict(
-	ctx context.Context,
-	reader client.Reader,
-	lease *coordinationv1.Lease,
-	matches func(*coordinationv1.Lease) bool,
-) (bool, error) {
-	if lease == nil || matches == nil {
-		return false, nil
-	}
-	latest := &coordinationv1.Lease{}
-	if err := reader.Get(ctx, types.NamespacedName{Namespace: lease.Namespace, Name: lease.Name}, latest); err != nil {
-		if apierrors.IsNotFound(err) {
-			return false, nil
-		}
-		return false, err
-	}
-	return matches(latest), nil
 }
 
 func (r *TaskReconciler) releaseSubstratePoolActorLeasesAfterTerminalCleanup(ctx context.Context, task *corev1alpha1.Task) (bool, error) {
@@ -1395,88 +1295,6 @@ func (r *TaskReconciler) taskHasSubstratePoolActorLeases(ctx context.Context, ta
 	return len(leases) > 0, nil
 }
 
-func newSubstratePoolActorLease(
-	task *corev1alpha1.Task,
-	namespace string,
-	name string,
-	actorID string,
-) *coordinationv1.Lease {
-	lease := &coordinationv1.Lease{
-		ObjectMeta: metav1.ObjectMeta{
-			Namespace: namespace,
-			Name:      name,
-		},
-	}
-	setSubstratePoolActorLeaseHolder(lease, task, actorID)
-	return lease
-}
-
-func setSubstratePoolActorLeaseHolder(lease *coordinationv1.Lease, task *corev1alpha1.Task, actorID string) {
-	if lease.Labels == nil {
-		lease.Labels = map[string]string{}
-	}
-	lease.Labels[labels.LabelManaged] = managedLabelValue
-	lease.Labels[labels.LabelPurpose] = substratePoolActorLeasePurpose
-	lease.Labels[substratePoolActorLeaseActorIDLabel] = labels.SelectorValue(actorID)
-	lease.Labels[substratePoolActorLeaseHolderUIDLabel] = labels.SelectorValue(string(task.UID))
-	if lease.Annotations == nil {
-		lease.Annotations = map[string]string{}
-	}
-	lease.Annotations[substratePoolActorLeaseTaskNSAnno] = task.Namespace
-	lease.Annotations[substratePoolActorLeaseTaskNameAnno] = task.Name
-	lease.Annotations[substratePoolActorLeaseTaskUIDAnno] = string(task.UID)
-	now := metav1.NewMicroTime(time.Now())
-	holder := fmt.Sprintf("%s/%s/%s", task.Namespace, task.Name, task.UID)
-	lease.Spec.HolderIdentity = &holder
-	lease.Spec.AcquireTime = &now
-	lease.Spec.RenewTime = &now
-}
-
-func substratePoolActorLeaseHeldByTask(lease *coordinationv1.Lease, task *corev1alpha1.Task) bool {
-	if lease == nil || task == nil || task.UID == "" || lease.Annotations == nil {
-		return false
-	}
-	return lease.Annotations[substratePoolActorLeaseTaskNSAnno] == task.Namespace &&
-		lease.Annotations[substratePoolActorLeaseTaskNameAnno] == task.Name &&
-		lease.Annotations[substratePoolActorLeaseTaskUIDAnno] == string(task.UID)
-}
-
-func substratePoolActorLeaseActorID(lease *coordinationv1.Lease) string {
-	if lease == nil {
-		return ""
-	}
-	if lease.Labels != nil {
-		if actorID := strings.TrimSpace(lease.Labels[substratePoolActorLeaseActorIDLabel]); actorID != "" {
-			return actorID
-		}
-	}
-	return strings.TrimSpace(lease.Name)
-}
-
-func substratePoolActorLeaseName(actorID string) string {
-	return strings.TrimSpace(actorID)
-}
-
-func substratePoolActorLeaseActiveTaskPhase(phase corev1alpha1.TaskPhase) bool {
-	switch phase {
-	case "", corev1alpha1.TaskPhasePending, corev1alpha1.TaskPhaseScheduled, corev1alpha1.TaskPhaseRunning:
-		return true
-	default:
-		return false
-	}
-}
-
-func taskExecutionWorkspaceCleanupSucceeded(task *corev1alpha1.Task) bool {
-	if task == nil {
-		return false
-	}
-	return statusrules.CleanupSucceeded(task.Status.ExecutionWorkspace)
-}
-
-func taskSubstratePoolActorCleanupRequired(task *corev1alpha1.Task) bool {
-	return task != nil && !taskExecutionWorkspaceCleanupSucceeded(task)
-}
-
 func taskExecutionWorkspacePhase(task *corev1alpha1.Task) corev1alpha1.ExecutionWorkspacePhase {
 	if task == nil || task.Status.ExecutionWorkspace == nil {
 		return ""
@@ -1489,23 +1307,6 @@ func taskExecutionWorkspaceReason(task *corev1alpha1.Task) corev1alpha1.Executio
 		return ""
 	}
 	return task.Status.ExecutionWorkspace.Reason
-}
-
-func substratePoolActorOrdinalFromID(actorID string, prefix string) (int, bool) {
-	actorID = strings.TrimSpace(actorID)
-	prefix = strings.Trim(strings.TrimSpace(prefix), "-")
-	suffix, ok := strings.CutPrefix(actorID, prefix+"-")
-	if !ok || len(suffix) != 5 {
-		return 0, false
-	}
-	ordinal := 0
-	for _, ch := range suffix {
-		if ch < '0' || ch > '9' {
-			return 0, false
-		}
-		ordinal = ordinal*10 + int(ch-'0')
-	}
-	return ordinal, true
 }
 
 // handleRunning handles Tasks in Running phase
